add map and change int to float for temp

+++ src/main/java/it/edu/marconilatina/arduino/model/sensor/ISensor.java
@@ -3,6 +3,6 @@ package it.edu.marconilatina.arduino.model.sensor;
 public interface ISensor {
 
 	public String getSensorName();
-    public double getValue();
+    public float getValue();
     public long getTimestamp();
 }
